perf harness: fix unsigned-wrap in perf_fill_double (tpsv ill-conditioning)
`(v % 211u) - 105u` was unsigned arithmetic: every residue < 105 wrapped to
~1.8e19 and divided to ~8.7e16, so half the fill values were ~1e16 garbage
instead of rationals in [-0.5, 0.5]. Harmless for single-pass kernels
(matvec/axpy/dot — subject and migrated legs see identical inputs), but it
made triangular SOLVE inputs catastrophically ill-conditioned: the solution
overflows to inf over N steps and x87 inf-arithmetic inflates the timing ~100x.

This was the true cause of the etpsv/ytpsv N=512 "par4 > par1" cells previously
filed as an x87 extreme-value measurement caveat. Cast to signed before centring.

After-fix REPS=3 5-way sweep (cmp5_gapk10_fillfix_raw.tsv): all tpsv cells now
show clean par4/par1 in [0.42, 0.69], sane O(N^2) scaling, no million-ns blowups,
p1/mig ~= 1.00. etpsv LTU 512 par4/par1 1.869 -> 0.491; ytpsv LCN/LCU/LTN/LTU 512
~1.25-1.31 -> ~0.45.

+++ tests/epblas-parallel/perf/perf_common.h
@@ -124,9 +124,15 @@ static inline int perf_parse_int_list(const char *env_name,
  * BLAS calls. Same seed = same sequence; we reset before each timed
  * block. */
 static inline double perf_fill_double(size_t i, int salt) {
-    /* small rational in [-1, 1]. */
+    /* small rational in [-1, 1]. NOTE: the centring subtraction MUST be signed
+     * — `(v % 211u) - 105u` is unsigned, so every residue < 105 wraps to ~1.8e19
+     * and divides to ~8.7e16, injecting ~1e16 garbage into half the fill. Benign
+     * for single-pass kernels (matvec/axpy/dot: both legs see the same values),
+     * but it makes triangular SOLVE inputs catastrophically ill-conditioned, so
+     * the solution overflows over N steps and x87 inf-arithmetic inflates the
+     * timing ~100x. Cast to signed before centring. */
     size_t v = (i * 2654435761u) ^ (size_t)(salt * 0x9e3779b9u);
-    return (double)((v % 211u) - 105u) / 211.0;
+    return (double)((long)(v % 211u) - 105) / 211.0;
 }
 
 static inline void perf_print_header(void) {
